feat: Data flow using react context api
