+++ splunk_add_on_ucc_framework/ucc_ui_lib/src/main/webapp/components/table/Table.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useState, useContext } from 'react';
 import Table from '@splunk/react-ui/Table';
 import Switch from '@splunk/react-ui/Switch';
 import ButtonGroup from '@splunk/react-ui/ButtonGroup';
@@ -12,11 +12,15 @@ import PropTypes from 'prop-types';
 import { ActionButtonComponent } from './TableStyle';
 import { getUnifiedConfigs } from '../../util/util';
 import { getExpansionRow } from './TableExpansionRow';
+import InputRowContext from '../../context/InputRowContext';
 
 function InputTable({ isInput, serviceName, data, handleToggleActionClick }) {
 
     const [sortKey, setSortKey] = useState('name');
     const [sortDir, setSortDir] = useState('asc');
+
+    const { rowData, setRowData } = useContext(InputRowContext);
+    console.log("rowData: ", rowData);
 
     const generateColumns = () => {
         const unifiedConfigs = getUnifiedConfigs();
@@ -117,14 +121,14 @@ function InputTable({ isInput, serviceName, data, handleToggleActionClick }) {
                 key={row.id}
                 expansionRow={getExpansionRow(columns.length, row)}
             >
-                {columns && columns.length && columns.map((header) => {
+                {columns && columns.length && columns.map((header, index) => {
                     if (header.field === "disabled") {
                         return (
                             <Table.Cell key={header.field}>
                                 <Switch
                                     key={row.name}
                                     value={row.disabled}
-                                    onClick={() => handleToggleActionClick(row)}
+                                    onClick={() => handleToggleActionClick(row, index)}
                                     selected={!row.disabled}
                                     appearance="toggle"
                                 >
