+++ splunk_add_on_ucc_framework/ucc_ui_lib/src/main/webapp/context/InputRowContext.jsx
@@ -1,0 +1,19 @@
+import React, { createContext, useState } from "react";
+
+const InputRowContext = createContext({
+    rowData: {},
+    setRowData: () => { },
+});
+
+export const InputRowContextProvider = ({ children }) => {
+
+    const [rowData, setRowData] = useState({});
+
+    return (
+        <InputRowContext.Provider value={{ rowData, setRowData }}>
+            {children}
+        </InputRowContext.Provider>
+    );
+}
+
+export default InputRowContext;
