+++ splunk_add_on_ucc_framework/ucc_ui_lib/src/main/webapp/pages/Input/InputPageStyle.jsx
@@ -3,10 +3,12 @@ import { variables } from '@splunk/themes';
 
 export const TitleComponent = styled.div`
     font-size: ${variables.fontSizeXXLarge};
+    margin-bottom: "20px";
 `;
 
 export const SubTitleComponent = styled.div`
     font-size: ${variables.fontSize};
+    margin-bottom: "10px";
 `;
 
 export const TableCaptionComponent = styled.div`
